% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_core.R
\name{print.toon}
\alias{print.toon}
\title{Print method for toon class}
\usage{
\method{print}{toon}(x, ...)
}
\arguments{
\item{x}{TOON string object}

\item{...}{Additional arguments (ignored)}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print method for toon class
}
\examples{
x <- to_toon(list(a = 1, b = 2))
print(x)

}
