% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_toonlite_parse}
\alias{stop_toonlite_parse}
\title{Signal a toonlite parse error}
\usage{
stop_toonlite_parse(
  message,
  line = NULL,
  column = NULL,
  snippet = NULL,
  file = NULL
)
}
\arguments{
\item{message}{Error message}

\item{line}{Line number}

\item{column}{Column number}

\item{snippet}{Code snippet}

\item{file}{File path}
}
\description{
Signal a toonlite parse error
}
\keyword{internal}
