% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{warn_toonlite}
\alias{warn_toonlite}
\title{Signal a toonlite warning}
\usage{
warn_toonlite(message, type = "general", details = NULL)
}
\arguments{
\item{message}{Warning message}

\item{type}{Warning type}

\item{details}{Additional details}
}
\description{
Signal a toonlite warning
}
\keyword{internal}
